ci: Run clang-tidy in github actions linter
Adds a github action workflow to run clang-tidy - should have better integration with github,
using file annotations to mark issues directly in the PR.

Build dependencies need to be installed to get cmake to generate the compilation database.
This is now using the packaging as source of truth instead of hardcoding the list.

The externals don't really need to be compiled to get the compilation db, but I added caching (for the externals and the build-dep metapackage), so now that doesn't take much time.